refactor: 파일 이름 컨벤션에 맞게 수정 (#60)
