+++ app/(header)/club/schedule/ScheduleOneGameInfo.tsx
