+++ app/(header)/club/schedule/ScheduleOneGameResult.tsx
